+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.17.31-SNAPSHOT
+### is.codion.swing.framework.server.monitor
+- ServerMonitor and ServerMonitorPanel, tables migrated from plain JTable to FilteredTable.
 
 ## 0.17.30
 ### is.codion.common